[dotnet pack] does not handle semantic versions properly.  It removes leading zeros from any embedded versions numbers, so "18.09" becomes "18.9" when the package ius created.
This is an issue now because I decided the use the date based versioning scheme for the library packages.

It might be possible to fix this by using NuGet to build and package the project instead of dotnet, but I don't want to take the time to debug this now.

So, we're dropping the leading "0" in the month part of the version.