[provisoning]: Wait 5 seconds before growing the VM primary partition for Hyper-V and XEN.
I just saw a transient failure when executing [growpart] this morning when deploying to XEN.  I'm guessing that this happened because we just rebooted the node and perhaps the filesystem isn't quite ready for this operation.

I'm going to add a 5 second delay for both Hyper-V and XEN just to be safe.