[build]: Hacked [nshell.csproj] to get [dotnet publish] to work.
I had to remove the "s" in "<RuntimeIdentifiers>" to "<RuntimeIdentifier>".  I'm not sure why I needed to do this because other projects like [neon-cli] work with this with the "s".

I'm going with this fix for now and hopefully this will be addressed with the shipping version of VS 2019 in a month or so.  This won't matter until we need to build for multiple platforms.