[revert]: [CadenceSettings.ServerUris] back to [Servers].  Here's the because-why:
neonKUBE uses this convention in other places like [CouchbaseSettings] and I'm trying to use consistent naming conventions across the entire neonKUBE library suite.  The goal here is for neonKUBE to eventually become a popular way for .NET developers (and potentially others) to do major Kubernetes-based projects and consistency helps (or at least can't hurt).

In the distant past (maybe 2-3 years ago I think I had originally called this [ServerUris] for the Couchbase settings and then decided it was a bit too wordy and changed it to just [Servers].