[openebs]: removed the Default and HostPath engines types and make Jiva the default.  Default used to resolve to Jiva for small clusters and cStor for large clusters.  We no longer do this.
The  code also used to install Jiva always and then layer cStor on top when selected as the engine.  We also ignored the HostPath engine when specified and installed Jiva regardless.

This is pretty confusing since the storage engine property in the cluster definion implies that only the specified engine will be required.

I don't think it really makes sense for both cStor and Jiva engines to be deployed at the same time, so I'm changing that.