Removed exclusion tags and CleanCrds target from Neon.Kube.Setup.csproj
In this commit, several lines of code were removed from the `Neon.Kube.Setup.csproj` file. These lines primarily consisted of `<None Remove="..."/>` tags, which are used to exclude specific files or directories from the project. The removal of these lines indicates that these files or directories are no longer being excluded from the project. Additionally, a block of code related to the "CleanCrds" target was also removed. This target was used to delete specific files before the "Clean" target is run. The removal of this block suggests that these files are no longer being deleted as part of the build process.